refactor: 分离每个芯片的实现
