+++ sftool-lib/src/sf32lb56/write_flash.rs
@@ -1,0 +1,7 @@
+use crate::write_flash::WriteFlashTrait;
+
+impl WriteFlashTrait for super::SF32LB56Tool {
+    fn write_flash(&mut self) -> Result<(), std::io::Error> {
+        todo!("SF32LB56 write_flash implementation not yet available")
+    }
+}
