+++ sftool-lib/src/sf32lb58/write_flash.rs
@@ -1,0 +1,7 @@
+use crate::write_flash::WriteFlashTrait;
+
+impl WriteFlashTrait for super::SF32LB58Tool {
+    fn write_flash(&mut self) -> Result<(), std::io::Error> {
+        todo!("SF32LB58 write_flash implementation not yet available")
+    }
+}
